requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.3.0
certifi
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.0
decorator==5.1.1
entrypoints==0.4
ffmpeg-python==0.2.0
filelock==3.10.0
frozenlist==1.3.3
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
lit==15.0.7
llvmlite==0.39.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.2
openai-whisper @ git+https://github.com/openai/whisper.git@6dea21fd7f7253bfe450f1e2512a0fe47ee2d258
packaging==23.0
panda==0.3.1
pandas==1.5.3
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytube @ git+https://github.com/pytube/pytube@d3d18691b3e99b2d3b4d620446b088a1c32896c6
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.2
rich==13.3.2
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.15.0
sympy==1.11.1
tiktoken==0.3.1
toml==0.10.2
toolz==0.12.0
torch==2.0.0
tornado==6.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
yarl==1.8.2
zipp==3.15.0
